Sorts response objects keys so that etag works
Because JSON.stringify outputs objects attributes in an almost
random order (depends on the order of insertion of the object attributes),
the etag would always change. If we add the objects keys (add the
keys to the obj) in a specific order (the sorted order in this case),
then the JSON output will always be the same, and then the etag
will work correctly.