Castration emails feature
Two endpoints added:
- /checkCastrationEmails: checks all pets to see which needs a castration
email sent, and sends those emails with a link that the receiver can use
to postpone the castration date by 15 days (configurable). The link points
to the endpoint below.
- /:id/postponeCastrationDate: postpones the castration date by 15 days,
and is a GET although it modifies the data, because it is used as a link
in the e-mail (so it wouldn't be possible to do a PUT for example). The
link only works once, subsequent calls will result in an error.